Module quirc_impl::ffi

Structs

FfiQuirc

FfiQuircCode
This structure is used to return information about detected QR codes in the input image — FFI.

FfiQuircData
This structure holds the decoded QR-code data

FfiQuircPoint
This structure describes a location in the input image buffer.

Enums

QuircDecodeError
This enum describes the various decoder errors which may occur.

Functions

quirc_begin⚠
These functions are used to process images for QR-code recognition. quirc_begin() must first be called to obtain access to a buffer into which the input image should be placed. Optionally, the current width and height may be returned.

quirc_count⚠
Return the number of QR-codes identified in the last processed image.

quirc_decode⚠
Decode a QR-code, returning the payload data.

quirc_destroy⚠
Destroy a QR-code recognizer.

quirc_end⚠
After filling the buffer, quirc_end() should be called to process the image for QR-code recognition. The locations and content of each code may be obtained using accessor functions described below.

quirc_extract⚠
Extract the QR-code specified by the given index.

quirc_new
Construct a new QR-code recognizer.

quirc_resize⚠
Resize the QR-code recognizer. The size of an image must be specified before codes can be analyzed.

quirc_strerror
Return a string error message for an error code.

quirc_version
Obtain the library version string.