Struct quirc_impl::ffi::FfiQuircData

#[repr(C)]

pub struct FfiQuircData {
    pub version: c_int,
    pub ecc_level: c_int,
    pub mask: c_int,
    pub data_type: c_int,
    pub payload: [u8; 8896],
    pub payload_len: c_int,
    pub eci: u32,
}

This structure holds the decoded QR-code data

Fields
version: c_int

Various parameters of the QR-code. These can mostly be ignored if you only care about the data.

ecc_level: c_int mask: c_int data_type: c_int

This field is the highest-valued data type found in the QR code.

payload: [u8; 8896]

Data payload. For the Kanji datatype, payload is encoded as Shift-JIS. For all other datatypes, payload is ASCII text.

payload_len: c_int eci: u32

ECI assignment number