Struct quirc_impl::ops::QuircCode

pub struct QuircCode {
    pub corners: [QuircPoint; 4],
    pub size: u32,
    pub cell_bitmap: [u8; 3917],
}

This structure is used to return information about detected QR codes in the input image.

Fields
corners: [QuircPoint; 4]

The four corners of the QR-code, from top left, clockwise

size: u32

The number of cells across in the QR-code.

cell_bitmap: [u8; 3917]

The cell bitmap is a bitmask giving the actual values of cells.

If the cell at (x, y) is black, then the following bit is set:

cell_bitmap[i >> 3] & (1 << (i & 7))

where i = (y * size) + x.

Methods

impl QuircCode

pub fn decode(&self, data: &mut QuircData) -> QuircDecodeError

Decode a QR-code, returning the payload data