Struct quirc_impl::ffi::FfiQuircCode

#[repr(C)]

pub struct FfiQuircCode {
    pub corners: [FfiQuircPoint; 4],
    pub size: c_int,
    pub cell_bitmap: [u8; 3917],
}

This structure is used to return information about detected QR codes in the input image — FFI.

Fields
corners: [FfiQuircPoint; 4]

The four corners of the QR-code, from top left, clockwise

size: c_int

The number of cells across in the QR-code.

cell_bitmap: [u8; 3917]

The cell bitmap is a bitmask giving the actual values of cells.

If the cell at (x, y) is black, then the following bit is set:

cell_bitmap[i >> 3] & (1 << (i & 7))

where i = (y * size) + x.